wifi-deauth-detection/requirements.txt
scapy
python-dotenv
google-generativeai
